backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
mediapipe==0.10.14
opencv-python-headless==4.9.0.80
numpy==1.26.4
pydantic==2.7.1
python-multipart==0.0.9
face-recognition==1.3.0
deepface==0.0.93
tensorflow==2.15.1
google-genai>=0.8.0
python-dotenv>=1.0.1
PyJWT>=2.8.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
faster-whisper==1.0.3
TTS==0.22.0
